Present user with some well-known options.
The API does not seem to offer a straightforward
way to get a list of all weather stations,
although the information is available on the web.
It appears to require a rather considerable amount
of manual labor to copy that data into some sort
of lookup table, which I shall postpone for now.

Besides, for the general public, presenting
options with recognizable names seems preferable.

A more fancy approach would be to display a map
of Iceland, and allow the user to click on it to
specify latitude and longitude, from which we can
find the nearest weather station.

Another approach would be to assume the user is
only interested in the location where they are at
at the moment, and get that from their ip-address,
but that ignores the needs of travellers preparing
for a trip out of town, for example.

So I shall stick to the recognizable names in this
text-based application.