Call the API.
Use Icelandic as default language.
Also, limit the weather stations to Reykjavík
for now. The goal is to allow the user to specify
which stations to pull data from, but that must
wait until later.

We shall use the xml response, as that seems
relatively accessible. The csv comes as an
attachment for example, and would take some extra
steps to process.