Pick the fruits of the xml tree.
Use defusedxml to guard against xml exploits.
Not that we should expect such from the national
weather forecast service, but there's no good
reason not to use this here.

I filter the forecast info immediately on the
time range, as we are traversing the list anyway,
so I figured we might as well avoid multiple
passes as well as keeping more info in memory than
we strictly need.

But perhaps it would be cleaner to separate the
filtering logic from the extraction.

It's not like the list of predictions will ever be
huge, it only goes so far, maybe a couple of
weeks, with one hour intervals.

So memory is not an issue here, nor is making two
passes through the list.

Even sorting the whole list shouldn't be too
demanding, but it seems it isn't necessary anyway:
From some manual observations, the findall()
method seems to consistently preserve the order of
predictions from the xml, although I cannot find
any verification of that fact in the defusedxml
documentation.

In case there are no predictions that fall within
the desired interval, I also preserve the closest
measurement outside the interval on both sides
(if there are any at all), to be able to display
those as a backup.